rename: 공통 컴포넌트들을 상위로 이동
여러 step에서 재사용할 수 있도록 공통 컴포넌트들을 상위로 이동함
