+++ src/pages/createBill/addExpenseStep/index.tsx
@@ -5,7 +5,7 @@ import { Close } from '@/assets/svgs/icon';
 import Header from '@/common/components/Header';
 import { BaseFunnelStepComponentProps } from '@/common/types/useFunnel.type';
 import expense from '@/service/apis/expense';
-import FormCard from '@/pages/createBill/addExpenseStep/components/FormCard';
+import FormCard from '@/pages/createBill/components/FormCard';
 import useAddExpenseFormArray from '@/pages/createBill/hooks/useAddExpenseFormArray';
 import getTotalExpense from '@/pages/createBill/utils/getTotalExpense';
 import { BillContext } from '@/pages/createBill/types/billContext.type';
