+++ src/pages/createBill/components/NumPadBottomSheet/index.tsx
@@ -5,7 +5,7 @@ import {
   DrawerRoot,
   DrawerTrigger,
 } from '@/common/components/Drawer/drawer';
-import NumPad from '@/pages/createBill/addExpenseStep/components/NumPad';
+import NumPad from '@/pages/createBill/components/NumPad';
 import * as S from './index.styles';
 
 interface NumPadBottomSheetProps {
